# services/reference-service/src/main/resources/openapi/reference-api.yaml
...
openapi: 3.1.0
info:
  title: CardDemo Reference Service API
  version: 1.0.0
  summary: >-
    Reference data of the migrated CardDemo credit-card management application:
    transaction types and their categories, disclosure-group interest rates, the
    three United States address lookups, and date evaluation.
  description: >-
    The reference context owns the data every other context reads and none of
    them writes. It publishes nineteen operations, across thirteen paths, over
    six tables.


    Transaction types and transaction categories are maintained here, and the
    relationship between them is the one referential rule the baseline states
    anywhere in this domain: a category references its type, and the type cannot
    be deleted while a category still references it. That refusal is a
    first-class user outcome in the baseline rather than a fault, and it is
    reported here as 409 for the reason recorded on the Conflict response.


    Disclosure-group rows carry the interest rate the nightly interest
    calculation consumes. That rate is an arithmetic operand, not a display
    value, so it is transported as a JSON string for the reason recorded on the
    InterestRate schema, and the lookup operation documents the DEFAULT-group
    fallback the baseline performs when an account's own group has no matching
    row.


    The three address lookups -- phone area codes, state codes and state and
...
    operation that writes them.


...


...
    one field.
...

...
servers:
...
security:
  - bearerAuth: []

# WHY : Refactoring Rationale: the authority each operation requires is declared
#       as a machine-readable field on the operation itself, and not only in
#       prose and tag membership. Prose and tag membership are not checkable: a
#       reviewer can read that a write is administrative, and a generator or a
#       gateway cannot, so the restriction would be indistinguishable from any
#       other authenticated route. Every operation below carries
#       x-required-authority.
# WHY : Assumptions: the value on each operation is not chosen here. It is
#       derived from what SecurityConfig.java already enforces for this service:
#       POST, PUT, PATCH and DELETE on every path each require
#       JwtRoleConverter.ADMIN_AUTHORITY, at its lines 99 to 106, and every
#       remaining request needs only authentication, at line 107. So every write
#       below carries carddemo-admin and every READ carries the third value
#       authenticated, and the document describes the filter chain rather than a
#       second opinion about it.
# WHY : Refactoring Rationale: an earlier revision carried carddemo-user on the
#       reads, which stated something the filter chain does not enforce and could
#       not be enforced without a behavioural regression. The shared converter
#       grants exactly the groups a token presents and builds NO hierarchy -- its
#       conversion loop admits a group only if it is one of the two recognised
#       names -- so a token carrying carddemo-admin alone receives ADMIN_AUTHORITY
#       and nothing else. A gateway or generator enforcing carddemo-user on the
#       reads would therefore refuse every administrator, and the baseline refuses
#       an administrator nothing: app/cbl/COSGN00C.cbl sends type 'A' to the
#       administrative menu at its line 245 and every function is reachable from
#       there. The honest declaration is the one the chain makes.
# WHY : Alternatives Considered: granting USER_AUTHORITY to an administrator as
#       well, so that carddemo-user could be declared and enforced on the reads.
#       Rejected because the grant would have to be made in the shared converter,
#       where it would change the authority set every one of the nine contexts
#       sees, to buy nothing this service needs: a read that any authenticated
#       caller may perform is described exactly by saying so. The two GROUP values
#       remain exactly the two the shared converter recognises, ADMIN_AUTHORITY
#       and USER_AUTHORITY at
#       services/common-lib/src/main/java/com/carddemo/common/security/JwtRoleConverter.java
#       lines 250 and 260, which are the migrated form of the two SEC-USR-TYPE
#       values 'A' and 'U'; authenticated names a REQUIREMENT rather than a group,
#       and the model below distinguishes the two.
...
paths:
...
  /api/v1/reference/us-phone-area-codes:

    description: >-
      The seeded list of North American Numbering Plan area codes this platform
      accepts, each carrying the classification that says which of the two
      baseline sub-lists it belongs to. Read-only: this context seeds the list
      and account-service queries it during address maintenance, so there is no
      operation here that writes it.

    get:
      tags:
        - reference-data
      operationId: listUsPhoneAreaCodes
      x-required-authority: authenticated
      summary: List accepted phone area codes by key, one page at a time
      description: >-
        Returns one page of area codes in ascending code order, optionally
        narrowed to a single classification. Four hundred and ninety codes are
        seeded by db/migration/V2__seed_reference.sql, carried across from the
        condition-name list at app/cpy/CSLKPCDY.cpy line 30, whose own comment at
        lines 26 to 28 records the numbering-plan administrator's report as their
        source.


        The classification narrows the list to one of the two sub-lists the
        copybook overlays on the same three-character field: four hundred and ten
        general-purpose codes from its line 521, and eighty easily-recognisable
        codes from its line 931. Omitting the filter returns the whole
        four-hundred-and-ninety-code list, which is the only membership test
        account-service performs.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PagingDirection'
        - $ref: '#/components/parameters/CodeClassFilter'
      responses:
        '200':
          description: >-
            One page of accepted area codes with their classifications.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsPhoneAreaCodePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/reference/us-phone-area-codes/{areaCd}:

    description: >-
      One accepted phone area code. Addressing a single code turns the
      membership test the baseline performs against a condition-name list into a
      primary-key probe, so a caller validating one value need not page the whole
      list to do it.

    parameters:
      - $ref: '#/components/parameters/AreaCdPath'

    get:
      tags:
        - reference-data
      operationId: getUsPhoneAreaCode
      x-required-authority: authenticated
      summary: Test whether one phone area code is accepted
      description: >-
        Returns the area code and its classification when the code is accepted,
        and 404 when it is not. The 404 is the answer rather than a failure:
        absence from this list is precisely what the baseline's condition-name
        test reports when it evaluates false, and it is the outcome
        account-service turns into a rejected address.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: >-
            The code is accepted; its classification is returned with it.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsPhoneAreaCode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/reference/us-states:

    description: >-
      The seeded list of accepted two-character state codes. Read-only, for the
      same reason as the area codes: this context seeds them and
      account-service reads them.

    get:
      tags:
        - reference-data
      operationId: listUsStates
      x-required-authority: authenticated
      summary: List accepted state codes by key, one page at a time
      description: >-
        Returns one page of accepted state codes in ascending order. Fifty-six
        codes are seeded by db/migration/V2__seed_reference.sql, carried across
        from the condition-name list at app/cpy/CSLKPCDY.cpy line 1013 over the
        two-character field declared on line 1012. The fifty-six are more than
        the states alone: the baseline's list also admits the district, the
        territories and the military mailing codes the postal service treats
        alike, so a caller must not expect fifty.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PagingDirection'
      responses:
        '200':
          description: >-
            One page of accepted state codes.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsStatePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/reference/us-states/{stateCd}:

    description: >-
      One accepted state code, addressed directly so that validating a single
      value is a key probe rather than a scan of the list.

    parameters:
      - $ref: '#/components/parameters/StateCdPath'

    get:
      tags:
        - reference-data
      operationId: getUsState
      x-required-authority: authenticated
      summary: Test whether one state code is accepted
      description: >-
        Returns the state code when it is accepted and 404 when it is not, with
        the 404 carrying the same meaning as for an area code: absence from the
        seeded list is the answer the baseline's membership test gives.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: >-
            The state code is accepted.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/reference/us-state-zip-prefixes:

    description: >-
      The seeded list of accepted state and postal-prefix combinations. Each
      entry is one four-character value rather than two fields, because the
      baseline builds the four bytes and tests the pair as a unit.

    get:
      tags:
        - reference-data
      operationId: listUsStateZipPrefixes
      x-required-authority: authenticated
      summary: List accepted state and postal-prefix combinations
      description: >-
        Returns one page of accepted combinations in ascending order. Two hundred
        and forty are seeded by db/migration/V2__seed_reference.sql, carried
        across from the condition-name list at app/cpy/CSLKPCDY.cpy line 1073
        over the four-character field declared on line 1072, itself subordinate
        to the group on line 1071.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PagingDirection'
      responses:
        '200':
          description: >-
            One page of accepted state and postal-prefix combinations.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsStateZipPrefixPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/reference/us-state-zip-prefixes/{stateZipCd}:

    description: >-
      One accepted state and postal-prefix combination, addressed as the single
      concatenated value the baseline tests.

    parameters:
      - $ref: '#/components/parameters/StateZipCdPath'

    get:
      tags:
        - reference-data
      operationId: getUsStateZipPrefix
      x-required-authority: authenticated
      summary: Test whether one state and postal-prefix combination is accepted
      description: >-
        Returns the combination when it is accepted and 404 when it is not. The
        two leading postal digits are significant and are compared as characters,
        so a combination whose postal half begins with a zero is a distinct value
        from one that does not.
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: >-
            The combination is accepted.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsStateZipPrefix'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        A JSON Web Token issued by the configured identity provider and validated
        as a resource-server token. Its cognito:groups claim is converted to
        authorities by
        services/common-lib/src/main/java/com/carddemo/common/security/JwtRoleConverter.java,
        where the baseline user type 'A' becomes carddemo-admin and 'U' becomes
        carddemo-user. Presenting no usable token yields 401; presenting a valid
        token that lacks the authority an operation requires yields 403.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: >-
        The paging position to continue from: copy the lastKey member of the page
        in hand to advance past it, or the firstKey member to retreat from it, and
        state the matching direction. Omit it, and the direction, to obtain the
        opening page. The value is opaque and must be echoed exactly as received;
        it is not a type code, a category code or a row ordinal, and it cannot be
        constructed, parsed or compared by a client. A cursor that cannot be
        opened, or one replayed against a different filter, a different collection
        or the direction it was not minted for, is reported as 400 rather than
        answered with the wrong page.
      schema:
        $ref: '#/components/schemas/CursorToken'
    PagingDirection:
      name: direction
      in: query
      required: false
      description: >-
        Which way to move from the supplied cursor. It is meaningful only
        alongside a cursor, and the two are sent together or not at all. Reading
        next takes rows from the cursor's position forward in ascending key order;
        reading previous takes rows before it in descending key order and presents
        them ascending, which is the migrated form of the backward cursor at
        app/app-transaction-type-db2/cbl/COTRTLIC.cbl lines 354 to 368. The two
        differ in whether the cursor's own position is included, as recorded on
        the cursor parameter.
      schema:
        $ref: '#/components/schemas/PageDirection'
...
    CodeClassFilter:
      name: codeClass
      in: query
      required: false
      description: >-
        Narrows the page to one of the two phone-area-code classifications.
        Omitting it returns the whole seeded list, which is the membership test
        account-service actually performs.
      schema:
        $ref: '#/components/schemas/PhoneAreaCodeClass'
...
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >-
        Identity of the unit of work this request belongs to, propagated into the
        service log record and echoed on the response. Supply a value to tie this
        request to work already in progress upstream; omit it and the service
        mints one. A value that does not fit the published bound and character set
        is refused with 400 before any handler runs. A value whose shape is that of a
        primary account number -- only digits and the three separators, with
        thirteen or more digits -- is refused on the same terms, because this
        identity is echoed and logged. The refusal carries the shared problem body
        and the response correlation header, the same as any other 400, and it names
        the rule it applied without ever repeating the value it refused.
      schema:
        $ref: '#/components/schemas/CorrelationId'
  headers:
    XCorrelationId:
      description: >-
        Identity of the unit of work that produced this response. Echoed from the
        request where the caller supplied a conforming value, and minted by the
        service where it did not. Quote it when reporting a failure, because it is
        the route from this response to the operational record.
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'

    # WHY : Refactoring Rationale: this header is declared because the service
    #       actually sends it and an earlier revision of this document did not say
    #       so. The 401 is rendered by the shared handler
    #       common/error/ApiErrorSecurityHandlers.UnauthenticatedEntryPoint, which
    #       delegates to the framework's BearerTokenAuthenticationEntryPoint FIRST
    #       and only then writes the body, precisely so this challenge is composed
    #       exactly as the OAuth 2.0 bearer-token specification requires. A client's
    #       token-refresh logic reads this header, so leaving it undeclared left a
    #       generated client with no typed access to the one header it needs on the
    #       one response it needs it on.
    # WHY : Assumptions: the value is not enumerated and no pattern is written for
    #       it. The framework composes it from the error it detected -- a bare
    #       scheme when no token was presented, and a scheme with an error code and
    #       description when one was presented and rejected -- so pinning a shape
    #       here would either be wrong for one of those two cases or would have to
    #       be widened until it constrained nothing.
...
  responses:
...
    Unauthorized:
      description: >-
        No usable bearer token was presented: the header was absent, the token was
        malformed, its signature or issuer did not validate, or it had expired.
        This is distinct from 403 and from 409, and the distinction is not
        cosmetic -- a caller acts on 401 by obtaining a token, on 403 by obtaining
        a different one, and on 409 by changing its request. No body member names
        the group the caller lacks, because at this point no verified identity
        exists to compare against one. The body carries code CARDDEMO-0401 and the
        sentence "Sign on to continue", identically on all four of those causes,
        alongside the bearer-token challenge below.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
        WWW-Authenticate:
          $ref: '#/components/headers/WwwAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...